+++ lib/ts/claims/primitiveClaim.ts
@@ -1,13 +1,13 @@
 import { PrimitiveClaim as PrimitiveClaimWebJS } from "supertokens-web-js/recipe/session";
 
-import type { ValidationFailureCallback, ValidationSuccessCallback } from "../types";
+import type { ValidationFailureCallback } from "../types";
 import type { PrimitiveClaimConfig } from "supertokens-web-js/recipe/session";
 
 export class PrimitiveClaim<T> extends PrimitiveClaimWebJS<T> {
     constructor(
         config: PrimitiveClaimConfig & {
-            onSuccessRedirection?: ValidationSuccessCallback;
             onFailureRedirection?: ValidationFailureCallback;
+            showAccessDeniedOnFailure?: boolean;
         }
     ) {
         super(config);
@@ -21,8 +21,8 @@ export class PrimitiveClaim<T> extends PrimitiveClaimWebJS<T> {
             ) => {
                 return {
                     ...validator(...args),
-                    onSuccessRedirection: config.onSuccessRedirection,
                     onFailureRedirection: config.onFailureRedirection,
+                    showAccessDeniedOnFailure: config.showAccessDeniedOnFailure,
                 };
             };
         }