+++ lib/ts/types.ts
@@ -28,16 +28,12 @@ export type GetRedirectionURLContext = {
     showSignIn?: boolean;
 };
 
-export type ValidationSuccessCallback =
-    | (({ userContext }: { userContext: any }) => Promise<string | undefined> | string | undefined)
-    | undefined;
-
 export type ValidationFailureCallback =
     | (({ userContext, reason }: { userContext: any; reason: any }) => Promise<string | undefined> | string | undefined)
     | undefined;
 
 export type SessionClaimValidator = SessionClaimValidatorWebJS & {
-    onSuccessRedirection?: ValidationSuccessCallback;
+    showAccessDeniedOnFailure?: boolean;
     onFailureRedirection?: ValidationFailureCallback;
 };